+++ system/src/Grav/Common/Page/Pages.php
@@ -88,6 +88,9 @@ class Pages
      */
     protected $ignore_hidden;
 
+    /** @var string */
+    protected $check_method;
+
     /**
      * @var Types
      */
@@ -226,6 +229,11 @@ class Pages
         return $this->baseUrl($lang, $absolute) . Uri::filterPath($route);
     }
 
+    public function setCheckMethod($method)
+    {
+        $this->check_method = strtolower($method);
+    }
+
     /**
      * Class initialization. Must be called before using this class.
      */
@@ -239,6 +247,10 @@ class Pages
         $this->instances = [];
         $this->children = [];
         $this->routes = [];
+
+        if (!$this->check_method) {
+            $this->setCheckMethod($config->get('system.cache.check.method', 'file'));
+        }
 
         $this->buildPages();
     }
@@ -947,7 +959,7 @@ class Pages
             $taxonomy = $this->grav['taxonomy'];
 
             // how should we check for last modified? Default is by file
-            switch (strtolower($config->get('system.cache.check.method', 'file'))) {
+            switch ($this->check_method) {
                 case 'none':
                 case 'off':
                     $hash = 0;